Fix repo creation
If we have freshly created repo there won't be 'refs/heads/master' so
repo.head won't work and it raises 'Rugged::ReferenceError'

I might be cleaner to check if it exist, but then it's branch on every
run, because of first-run creation. By catching the exception we only
need to pay the cost on first run.